+++ lua/obsidian/types.lua
@@ -27,6 +27,7 @@
 ---@field workspaces obsidian.workspace.WorkspaceSpec[]
 ---@field log_level? integer
 ---@field notes_subdir? string
+---@field exclude_dir? string[]
 ---@field templates? obsidian.config.TemplateOpts
 ---@field new_notes_location? obsidian.config.NewNotesLocation
 ---@field note_id_func? (fun(title: string|?, path: obsidian.Path|?): string)|?
@@ -57,6 +58,7 @@
 ---@field workspaces obsidian.workspace.WorkspaceSpec[]
 ---@field log_level integer
 ---@field notes_subdir string|?
+---@field exclude_dir string[]
 ---@field templates obsidian.config.TemplateOpts
 ---@field new_notes_location obsidian.config.NewNotesLocation
 ---@field note_id_func (fun(id: string|?, path: obsidian.Path|?): string)