+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support `workspace/symbol` to search through note, note aliases and headings.
 - `actions.move_note` to move current note to a folder in the current workspace.
 - Add `padding_top` option to `Note.insert_text` for configuring blank lines inserted at the top of notes.
+- `:Obsidian help` has cmdline completion for wiki pages.
 
 ### Fixed